Fix socket on Linux hanging up when server is processing things.
I added a timeout, and a poll method to wait for data to arrive. (100 ms)

Basically this meant that a client would hang up early, and cause a broken pipe.
Now we wait a sensible amount of time for the server to reply and poll the socket.
If we get no reply then we know something is broken.

Enabled warnings, as highest level and they register as compiler errors too.